Refactor: DailyBackground
