+++ src/main/java/ticketaka/mtvs3_final_backend/redis/daily/background/domain/DailyBackground.java
@@ -1,4 +1,4 @@
-package ticketaka.mtvs3_final_backend.redis.daily.background;
+package ticketaka.mtvs3_final_backend.redis.daily.background.domain;
 
 import jakarta.persistence.Id;
 import lombok.*;
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@RedisHash(value = "draw_result", timeToLive = 24 * 60 * 60) // 60 * 15
+@RedisHash(value = "daily_background", timeToLive = 24 * 60 * 60) // 1일
 public class DailyBackground {
 
     @Id
@@ -19,9 +19,11 @@ public class DailyBackground {
     private LocalDate lastRefreshDate;
 
     @Builder
-    public DailyBackground(String id, Integer refreshCount, LocalDate lastRefreshDate) {
+    public DailyBackground(String id, LocalDate lastRefreshDate) {
         this.id = id;
-        this.refreshCount = refreshCount;
+        this.refreshCount = DAILY_BACKGROUND_GENERATION_LIMIT;
         this.lastRefreshDate = lastRefreshDate;
     }
+
+    private static final Integer DAILY_BACKGROUND_GENERATION_LIMIT = 2;
 }
