+++ src/main/java/ticketaka/mtvs3_final_backend/redis/daily/background/repository/DailyBackgroundRedisRepository.java
@@ -1,0 +1,8 @@
+package ticketaka.mtvs3_final_backend.redis.daily.background.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import ticketaka.mtvs3_final_backend.redis.daily.background.domain.DailyBackground;
+
+public interface DailyBackgroundRedisRepository extends CrudRepository<DailyBackground, String> {
+
+}
